Separate tests into logical sections
