+++ src/LibreCards.Tests/Gameplay/GameTests.cs
@@ -4,7 +4,7 @@ using LibreCards.Core.Persistence;
 using Moq;
 using Xunit;
 
-namespace LibreCards.Tests
+namespace LibreCards.Tests.Gameplay
 {
     public class GameTests
     {
